requirements.txt
# API & web
fastapi
uvicorn[standard]

# Database
sqlalchemy
psycopg2-binary

# Config & validation
pydantic
python-dotenv

# HTTP
requests

# ML / NLP (Qwen, finetuning, embeddings)
torch
transformers
peft
accelerate
bitsandbytes
safetensors
sentencepiece
datasets
trl
numpy
sentence-transformers

# Redis & data
redis
polars
